+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ### BUG FIXES
 * Fixed bug in `plotAllometry` where `crossprod` was used instead of `tcrossprod`
+* Fixed typp causing error in `phylo.integration`
 
 # CHANGES IN GEOMORPH VERSION 3.3.2 (Minor Release)